requirements.txt
langchain-google-genai==0.1.0
langchain==0.1.0
faiss-cpu==1.7.1
pypdf==3.6.0
sentence-transformers==2.2.0
langchain-community==0.1.0
pandas==2.0.0
PyMuPDF==1.22.0
python-docx==0.8.11
SpeechRecognition==3.8.1
gTTS==2.2.3
pyaudio==0.2.11
pygame==2.5.0
transformers==4.32.0
streamlit==1.18.1
